+++ src-radar/models/quadrant.js
