+++ src-radar/stylesheets/_tip.scss
