+++ src-radar/util/ringCalculator.js
